requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.1.0
psycopg-pool>=3.2.0
yfinance>=0.2.0
google-search-results>=2.4.2
wikipedia>=1.4.0
sqlalchemy>=2.0.0
alembic>=1.13.0
